Use consistent variable names across GSL modules
Mostly this standardizes the use of variable names in calls to GSL
functions.  We try to stay consistent with the GSL documentation,
while differentiating between Limits of integration (now $la, $lb)
and Parameters of a function (now $pa, $pb).

Also added some documentation to this effect, and updated the links
to point to specific GSL documentation.